feat(fleet): update data type in model & entity

+++ internal/modules/fleet/entities/fleet.go
@@ -3,8 +3,8 @@ package fleetentity
 type Fleet struct {
 	ID        uint64 `gorm:"primaryKey"`
 	VehicleId string `gorm:"column:vehicle_id"`
-	Latitude  int64  `gorm:"column:latitude"`
-	Logitude  int64  `gorm:"column:longitude"`
+	Latitude  float64  `gorm:"column:latitude"`
+	Logitude  float64  `gorm:"column:longitude"`
 	Timestamp int64  `gorm:"column:timestamp"`
 }
