+++ internal/modules/fleet/model/model.go
@@ -2,7 +2,7 @@ package fleetmodel
 
 type FleetResponse struct {
 	VehicleId string `json:"vehicle_id"`
-	Latitude  int64  `json:"latitude"`
-	Logitude  int64  `json:"longitude"`
+	Latitude  float64  `json:"latitude"`
+	Logitude  float64  `json:"longitude"`
 	Timestamp int64  `json:"timestamp"`
 }
